README.md File
The inspiration for doing this application was the desire to share cartoon enjoyment on a larger scale. 

Initially, it started as a humorous idea proposed to my friend, but it always felt as though it had potential, mainly because of how many possible versions of this simple idea are actually possible and can be implemented, given enough time. The target user for this application is the general public, though it is a bit skewed towards the younger audience, due to the desire for a simple interface. 

I am most proud of the application's ability to track finger touches and send similar data to other androids. It was a hectic experience to say the least, especially in a hackathon, to produce such a feat without android programming experience, but it was exciting to see the end results.